Rework capturing so that captures are decided during compilation
The new parser isn't able to look ahead to for loop arguments to make them
available to inline expressions for capturing.

e.g. i needs to be available for capturing in the inline function bodies.
multipliers = [(|n| n * i) for i in 1..=4]

The approach taken in this commit is to flag all non-local accesses in a
function as being potential captures (they could also be global accesses),
then during compilation we have enough information (by inspecting all locally
available value ids) to decide if a non-local access is a capture or not.