Only perform bytecode validation when debug assertions are enabled
The safety checks when reading bytecode have a continous runtime CPU cost of
2-4%, and if we can assume that bytecode has come from the compiler, then the
bytecode will be valid.

If bytecode has come from an external source then all bets are off, so it seems
to me that we can be bold here and skip incremental validation.

I think that if validation turns out to be desirable in practice, then an
improvement to this could be to have a validation step on any bytecode that is
to be read by an instruction reader:
- Read through the bytecode, checking that each opcode has the correct number of
  following bytes.
- Cache the addresses of each validated opcode
- Read through the bytecode again, and check that each jump op has a target
  that matches a validated opcode.