Remove support for setting Num2/Num4 elements by index
It was a mistake to add this feature; only values with shared underlying data
should be treated as mutable.

The following example demonstrates an issue with the current implementation:

```
foo = {x: make_num2 1, 2}
foo.x[0] = 99
foo.x
```

It would likely be possible to add support for the above by reassigning the
modified value to `foo.x`, but consider that it wouldn't be possible to write a
function like `num2.set 0, 99` that mutates the value in place (the function
wouldn't have access to the original value, only a copy).