Remove the 'assume valid bytecode' paths
Invalid bytecode can produce undefined behaviour, so valid bytecode reaaally
shouldn't be assumed by default.

The runtime overhead of bounds checking is currently in the range of 0-1%,
so it doesn't seem like it's worth keeping this logic around
(it could be reintroduced in the future for usecases where valid bytecode can
be assumed with total certainty).