Remove the docs symlink in core/cli
Symlinks aren't enabled by default for Git on Windows, and require admin
privileges to enable, which makes it impossible to build the CLI on Windows
machines without admin privileges.

The simple fix for now is to use relative paths when importing the docs for the
CLI's help system, although this will break release packaging. A solution can be
found when it's time to make the next release.