Make Koto thread-safe by default
Koto was previously thread-safe and had some parallel utilities in the core
library, and was made single-threaded for the performance benefits (5-15% faster
in benchmarks) and to simplify the implementation. The assumption was that
non-core modules could introduce parallel utilities later.

In hindsight this was a mistake because if Koto values aren't Send+Sync then
they can't be handed over to workers, forcing serialization at thread
boundaries.

I think it makes sense now to make Koto thread-safe by default, and now that the
koto_memory abstractions are in place it's easy enough to allow users to choose
the single-threaded variant if it makes sense for their application.