Fix chained compound assignment operations
Compound assignment operations need to be right associative for them to make
sense:

e.g.

in:

a, b, c = 1, 2, 3
a += b += c

left associative would imply
(a += b) += c

which results in `+= c` being applied to a temporary.

The correct order of things is:
a += (b += c)

Additionally, all compound assignment ops should share the same precedence to
ensure that this left-to-right ordering doesn't get broken.