Remove the `$` prefix and id-only variant from string interpolation
Now that rust-style formatting has been added to string interpolation,
it seems more natural to follow rust formatting as closely as possible.

It was also starting to bother me that there were two ways to render an
identifier (as $id or ${id}), with the id-only variant being less useful given
that it doesn't allow for format strings, so I think it's OK to standardize on
the `{}` syntax for all cases.