Fix packaging koto_cli
The docs need to be included in the crates/cli dir to allow cargo package to
find them. The top-level docs folder can be a symlink to the CLI docs dir.
Previously the symlink was the other way around (from the CLI dir to the
top-level dir), but that makes it harder to build on restricted machines
(see #264).