Prevent exhausted generators causing a panic
Calling `next()` on a generator with no more values to yield should
always return `null`, in line with what happens for normal iterators.

Prior to this patch, attempting to consume a value from an exhausted
generator (that is, one for which `next()` has already returned `null`)
would result in a runtime panic due to the call stack being empty when
invoking `execute_instructions()` from within `continue_running()`.