Remove support for generic type hints in the parser
For type hints to be practically useful for runtime checks, for now they're
going to have to be non-generic to avoid having to worry about (e.g.) checking
every element in a list to see if it matches `List<String>`.

In the future perhaps some strategies will be figured out for efficient runtime
checks, but for now we can simplify the parser and AST by making it clear that
types can't be nested.