Reserve sufficient registers for a frame when it's entered
The NewFrame op is run at the start of each frame, and tells the VM the
number of registers required by the frame's bytecode.
This results in allowing the stack size check in set_register to be
removed.

The VM may use additional temporary registers but will ensure that at
least the required registers are present in the stack.

A fair amount of reworking of the VM's call semantics has been gone in
to this, hopefully simplifying the logic and clarifying how frame setup
should be performed.