Add more `PartialEq` and `PartialOrd` impls for `KString` (#411)
Backport of 97ffe7afca181275b807c4f01d58d9c64fcbb75f

This avoids relying on automatic dereferencing to `str` in operator
arguments, which can fail if more impls become available.

In particular, this fixes an issue in the current `impl PartialOrd for
ValueKey`, where the comparison for the `Str` case fails if more impls
for strings become available.