Fix overflows when using iterators that output pairs in a for loop
This fixes a regression introduced in
45d3328661dba46dc8fea51cd53ef5705422a0f7 that caused register overflows
when using a for loop with iterators that output value pairs.

The idea in 45d33 was to push temporary registers onto the stack to
avoid assuming that it was safe to assign directly in to the stack.
This doesn't work with temporary tuples however because there's currently
no mechanism to clean up the pushed registers later on, which caused
the stack to grow rapidly.

Rather than introduce a new cleanup mechanism, here we roll back to the
previous approach but ensure that the the stack has available slots for
the temporary tuple contents. This might not be sufficient in the long
run but at the moment I can't think of a case where this won't hold up,
so it'll do for now.