Ensure that compound assignments in parentheses are evaluated correctly
Chains on the LHS of a compound assignment need to place their output in
the assigned result register. When the result of the assignment is
temporary (like when it's in parentheses on the LHS of a comparison),
the chain was assigning its own temporary output register rather than
using the result register assigned by the comparison op.