Add a feature to return the incomplete AST on parser error (#448)
This allows tools like `koto-ls` to provide information about the AST
up to the point where the error occurred.

This should be seen as a starting point that establishes a pattern
for providing as much of the incomplete AST as possible.

Some effort is made to repair the incomplete AST so that it can be
traversed without too many unattached nodes (functions and blocks are
repaired, along with the main block), but there are probably still
some cases where reattaching nodes would be useful.