Rename Parser::parse_destructure -> parse_binding
DestructureContext also gets renamed here to BindingContext,
and the Map variant is removed given that it only existed to prevent
type hints from being used in assignments to non-let destructured maps,
which would trigger an error anyway. The error gets renamed and reworded
here to give a hint to users that the should use `let` to enable type
hints.