Allow ignored unpacked values to have type hints
Ignored values don't get allocated registers, so a temporary register
is used to unpack the value when its type needs to be checked.

The previous approach to calling functions with a temporary tuple assumed
that the temp tuple's contents could be placed in the first temp
registers in the frame, so unpacking the ignored value into a temp
register overwrote the temp tuple.

The fix here is to place the temp tuple's contents in the registers
immediately preceding the frame (which is possible now that temp tuples
use an absolute index for their start register), so the type checking
is free to make use of the temp registers.